DRY with conftest
