+++ ckanext/geodatagov/tests/test_datajson.py
@@ -1,45 +1,23 @@
 import json
-import os
 import pytest
+import logging
 
-from ckan.tests.helpers import reset_db
 from ckan.tests.factories import Organization
 from ckan import model
-from ckan.model.meta import Session, metadata
 from factories import (DataJsonHarvestSourceObj,
                        HarvestJobObj)
 import ckanext.harvest.model as harvest_model
 from ckanext.datajson.harvester_datajson import DataJsonHarvester
-import mock_static_file_server
-import logging
 
+from utils import PORT, populate_locations_table
 
 log = logging.getLogger(__name__)
 
 @pytest.mark.usefixtures("with_plugins")
 class TestDataJsonHarvester(object):
 
-    @classmethod
-    def setup_class(cls):
-        log.info('Starting mock http server')
-        mock_static_file_server.serve(port=8996)        
-
     def setup_method(self):
-        # https://github.com/ckan/ckan/issues/4764
-        # drop extension postgis so we can reset db
-        os.system("PGPASSWORD=ckan psql -h db -U ckan -d ckan -c 'drop extension IF EXISTS postgis cascade;'")
-        reset_db()
-        os.system("PGPASSWORD=ckan psql -h db -U ckan -d ckan -c 'create extension postgis;'")
-        # echo "Downloading locations table"
-        os.system("wget https://github.com/GSA/datagov-deploy/raw/71936f004be1882a506362670b82c710c64ef796/"
-                  "ansible/roles/software/ec2/ansible/files/locations.sql.gz -O /tmp/locations.sql.gz")
-        # echo "Creating locations table"
-        os.system("PGPASSWORD=ckan psql -h db -U ckan -d ckan -c 'DROP TABLE IF EXISTS locations;'")
-        os.system("gunzip -c /tmp/locations.sql.gz | PGPASSWORD=ckan psql -h db -U ckan -d ckan -v ON_ERROR_STOP=1")
-        # echo "Cleaning"
-        os.system("rm -f /tmp/locations.sql.gz")
-        # os.system("ckan -c test.ini db upgrade -p harvest")
-        metadata.create_all(bind=Session.bind)
+        populate_locations_table()
 
     def run_gather(self, url):
         source = DataJsonHarvestSourceObj(url=url, owner_org=self.organization['id'])
@@ -103,7 +81,7 @@ class TestDataJsonHarvester(object):
         self.organization = Organization()
 
         # testing with data from https://www.consumerfinance.gov/data.json
-        url = 'http://127.0.0.1:8996/sample5_data.json'
+        url = f'http://127.0.0.1:{PORT}/sample5_data.json'
         obj_ids = self.run_gather(url=url)
         assert len(obj_ids) == 2
         self.run_fetch()
