+++ ckanext/geodatagov/tests/test_fix_packages.py
@@ -1,10 +1,7 @@
 import pytest
 import logging
-import os
 
-from ckan.tests.helpers import reset_db
 from ckan.tests import factories
-from ckan.model.meta import Session, metadata
 
 
 log = logging.getLogger(__name__)
@@ -14,12 +11,6 @@ class TestFixPkg(object):
 
     @classmethod
     def setup_class(cls):
-        # https://github.com/ckan/ckan/issues/4764
-        # drop extension postgis so we can reset db
-        os.system("PGPASSWORD=ckan psql -h db -U ckan -d ckan -c 'drop extension IF EXISTS postgis cascade;'")
-        reset_db()
-        os.system("PGPASSWORD=ckan psql -h db -U ckan -d ckan -c 'create extension postgis;'")
-        metadata.create_all(bind=Session.bind)
         cls.organization = factories.Organization()
 
     def test_fix_tags(self):
