+++ ckanext/geodatagov/tests/test_waf_GMI.py
@@ -1,36 +1,22 @@
 import json
 import logging
+import pytest
 
 import ckanext.harvest.model as harvest_model
-import mock_static_file_server
 from ckan import model
 from ckanext.geodatagov.harvesters.base import GeoDataGovWAFHarvester
 from factories import HarvestJobObj, WafHarvestSourceObj
-
-from ckan.tests.helpers import reset_db
 from ckan.tests.factories import Organization
-import pytest
-import os
-from ckan.model.meta import Session, metadata
+
+from utils import PORT, reset_db_and_solr
 
 log = logging.getLogger(__name__)
 
 @pytest.mark.usefixtures("with_plugins")
 class TestWafHarvester(object):
 
-    @classmethod
-    def setup_class(cls):
-        log.info('Starting mock http server')
-        mock_static_file_server.serve()
-
     def setup_method(self):
-        # https://github.com/ckan/ckan/issues/4764
-        # drop extension postgis so we can reset db
-        os.system("PGPASSWORD=ckan psql -h db -U ckan -d ckan -c 'drop extension IF EXISTS postgis cascade;'")
-        reset_db()
-        os.system("PGPASSWORD=ckan psql -h db -U ckan -d ckan -c 'create extension postgis;'")
-        # os.system("ckan -c test.ini db upgrade -p harvest")
-        metadata.create_all(bind=Session.bind)
+        reset_db_and_solr()
 
         self.organization = Organization()
 
@@ -100,7 +86,7 @@ class TestWafHarvester(object):
 
     def get_datasets_from_waf_gmi_sample(self):
         """ harvest waf-gmi/ folder as waf source """
-        url = 'http://127.0.0.1:%s/waf-gmi/index.html' % mock_static_file_server.PORT
+        url = f'http://127.0.0.1:{PORT}/waf-gmi/index.html'
 
         self.config1 = '{"private_datasets": "false"}'
         self.run_gather(url=url, source_config=self.config1)
